*** WORK IN PROGRESS *** Support Zip64-format jar files in the loader

+++ spring-boot-tools/spring-boot-loader/src/main/java/org/springframework/boot/loader/jar/CentralDirectoryParser.java
@@ -60,7 +60,7 @@ class CentralDirectoryParser {
 				.getInputStream(ResourceAccess.ONCE);
 		try {
 			int dataOffset = 0;
-			for (int i = 0; i < endRecord.getNumberOfRecords(); i++) {
+			for (long i = 0; i < endRecord.getNumberOfRecords(); i++) {
 				CentralDirectoryFileHeader fileHeader = CentralDirectoryFileHeader
 						.fromInputStream(inputStream);
 				visitFileHeader(dataOffset, fileHeader);
